Create package.json
This will get you setup with typescript as a devDependency so we can run the tsc command in our scripts. This was the --watch flag passed in so that it will compile (generate .js out of .ts files) every time the file is changed. You can execute this watcher by running your build command

```
npm run build
```